First Draft of the Expense Tracker
This is the first draft of the expense tracker. It will be improved upon in the future. I intend to use Flask or Django to give it a good UI and to also include some form of storage as currently data is lost when the program is closed by the user.